refactor: download button for Streamlit
