+++ .gitignore
@@ -32,5 +32,4 @@ tests/
 # OS generated files
 .DS_Store
 Thumbs.db
-
-data/
+rag_data/
